+++ site/astro.config.ts
@@ -1,0 +1,24 @@
+import mdx from "@astrojs/mdx";
+import react from "@astrojs/react";
+import solid from "@astrojs/solid-js";
+import tailwindcss from "@tailwindcss/vite";
+import { defineConfig } from "astro/config";
+
+// https://astro.build/config
+export default defineConfig({
+  srcDir: "src",
+
+  vite: {
+    // @ts-expect-error
+    plugins: [tailwindcss()],
+    server: {
+      allowedHosts: [".local"],
+    },
+  },
+
+  integrations: [
+    mdx(),
+    react({ include: ["**/*.react.*", "../packages/*react*/**"] }),
+    solid({ include: ["**/*.solid.*", "../packages/*solid*/**"] }),
+  ],
+});