+++ site/src/content.config.ts
@@ -1,0 +1,26 @@
+import { existsSync } from "node:fs";
+import { join } from "node:path";
+import { title } from "node:process";
+import { defineCollection, z } from "astro:content";
+import { glob } from "astro/loaders";
+
+const examples = defineCollection({
+  loader: glob({
+    pattern: "**/index.mdx",
+    base: join(import.meta.dirname, "examples"),
+  }),
+  schema: z.object({
+    title: z.string().optional(),
+    frameworks: z.array(
+      z.union([
+        z.literal("preact"),
+        z.literal("react"),
+        z.literal("solid"),
+        z.literal("svelte"),
+        z.literal("vue"),
+      ]),
+    ),
+  }),
+});
+
+export const collections = { examples };